# rubocop: reduce string to symbol conversion for performance (define_singleton_method)