--- a/lib/fluent/test/base.rb
+++ b/lib/fluent/test/base.rb
@@ -32,7 +32,7 @@ module Fluent
             #   klass.dup is worse because its ancestors does NOT include original class name
             klass_name = klass.name
             klass = Class.new(klass)
-            klass.define_singleton_method("name") { klass_name }
+            klass.define_singleton_method(:name) { klass_name }
             klass.module_eval(&block)
           end
           @instance = klass.new